[Modify] Locate Portal-modal into Card
- Card의 자식으로 Portal-modal을 집어넣음
- 모달이 뜰 때 fetch로 해당 상품의 정보를 가져오는 방식으로 구현
    - 혹시라도 실시간으로 재고 상황이 달라질까 싶어서 매번 fetch 하는 형식으로 구현했는데, 좋은 방법인지는 모르겠음
- StyledLi 밖에서 setModalState를 해주어야 함
    - 그렇지 않으면 클릭하는 모든 부분이 StyledLi라서 클릭 이벤트가 자꾸 발생해 false로 바꿔도 다시 true가 되어 모달창이 꺼지지 않는 일 발생
- 현재 useEffect 안쓰고 그냥 fetch를 사용했고, 동작은 하는데, 좋은 방법처럼 느껴지지는 않음
    - 근데 useEffect안에 fetch를 사용하는 것이 원칙일까.. 좀 더 공부해볼 필요가 있는 것 같음
