+++ banchan/src/components/utils/Card.jsx
@@ -3,34 +3,76 @@ import theme from "../utils/styles/theme.js";
 import Price from "./Price";
 import Label from "./Label";
 import { CenterContainer, LabelList, StyledDescription, StyledTitle } from "./styles/common.jsx";
+import { useEffect, useState } from "react";
+import Modal from "./modal/Modal.jsx";
+import ModalCard from "./modal/ModalCard.jsx";
+import { URLS } from "../../utils/variables.js";
+import IconButton from "./button/IconButton.jsx";
 
 const mockImage = "https://recipe1.ezmember.co.kr/cache/recipe/2020/09/23/5e308abb30b00ecb9c1b9b398db5b4451.jpg";
 
 const Card = ({ product, cardSize, margin = 0, type, onModal }) => {
+  const [modalState, setModalState] = useState(false);
+  const [detail, setDetail] = useState({});
+
+  const requestProductDetailInfo = async () => {
+    const data = await fetch(URLS.base + `detail/${product.detail_hash}`);
+    const json = await data.json();
+    setDetail({ ...json.data });
+    setModalState(true);
+  };
+
   return (
-    <StyledLi cardSize={cardSize} margin={margin} onClick={() => onModal(product)}>
-      <StyledThumbnail>
-        <StyledHoverLayer cardSize={cardSize}>
-          <DeliveryTypeList>
-            {product.delivery_type.reduce((acc, val, idx, array) => {
-              acc.push(
-                <p>
-                  {val} {idx < array.length - 1 ? <Divider /> : ""}
-                </p>
-              );
-              return acc;
-            }, [])}
-          </DeliveryTypeList>
-        </StyledHoverLayer>
-        <StyledImg cardSize={cardSize} src={type === "베스트" ? mockImage : product.image} alt="card-image" />
-      </StyledThumbnail>
-      <StyledTitle>{product.title}</StyledTitle>
-      <StyledDescription>{product.description}</StyledDescription>
-      <Price product={product} />
-      <LabelList>{product.badge && product.badge.map((e) => <Label badgeName={e} />)}</LabelList>
-    </StyledLi>
+    <>
+      <StyledLi cardSize={cardSize} margin={margin} onClick={requestProductDetailInfo}>
+        <StyledThumbnail>
+          <StyledHoverLayer cardSize={cardSize}>
+            <DeliveryTypeList>
+              {product.delivery_type.reduce((acc, val, idx, array) => {
+                acc.push(
+                  <p>
+                    {val} {idx < array.length - 1 ? <Divider /> : ""}
+                  </p>
+                );
+                return acc;
+              }, [])}
+            </DeliveryTypeList>
+          </StyledHoverLayer>
+          <StyledImg cardSize={cardSize} src={type === "베스트" ? mockImage : product.image} alt="card-image" />
+        </StyledThumbnail>
+        <StyledTitle>{product.title}</StyledTitle>
+        <StyledDescription>{product.description}</StyledDescription>
+        <Price product={product} />
+        <LabelList>{product.badge && product.badge.map((e) => <Label badgeName={e} />)}</LabelList>
+      </StyledLi>
+      {modalState && (
+        <Modal>
+          <ModalBackground>
+            <ModalContainer>
+              <ModalCard product={{ ...product, ...detail }} />
+              <IconButton type="CLOSE" fn={() => setModalState(false)} />
+            </ModalContainer>
+          </ModalBackground>
+        </Modal>
+      )}
+    </>
   );
 };
+
+const ModalBackground = styled(CenterContainer)`
+  position: fixed;
+  top: 0;
+  left: 0;
+  background: rgba(0, 0, 0, 0.4);
+  width: 100%;
+  height: 100%;
+  z-index: 5;
+`;
+
+const ModalContainer = styled.div`
+  display: flex;
+  align-items: flex-start;
+`;
 
 const StyledLi = styled.li`
   width: ${(props) => props.cardSize};
